+++ common/src/main/java/me/wesley1808/servercore/common/commands/ServerCoreCommand.java
@@ -68,7 +68,7 @@ public class ServerCoreCommand {
     }
 
     private static int reload(CommandSourceStack source) {
-        boolean success = Config.reloadMainConfig();
+        boolean success = Config.reloadConfigs();
         if (success) {
             source.sendSuccess(() -> Component.literal("Config reloaded!").withStyle(ChatFormatting.GREEN), false);
         } else {